Slab (ridge) plot for model predictions

Slab (ridge) plot for model predictions

mc_obs_slab(...)

mc_model_slab(..., n_sample = NA, draw = "collapse")

Arguments

...

Augments passed to ggdist::stat_slab.

n_sample

The number of sample of draws to show in slab plot. Default NA. If n_sample is NA, then mc_model_slab() will use all draws from posterior distribution.

draw

How to show draws? Default "collapse". If draw is "collapse", then all draws are collapsed together to show in one slab plot; if draw is "group", then each draw is shown in an individual slab plot; if draw is "hops", then mc_model_slab() will use animation to show each draw in one frame; if draw is an function, then all draws are aggregated by draw(). See examples for more details.